pyproject.toml
[tool.poetry]
name = "django-serials"
version = "0.1.0"
description = ""
authors = ["Ilya <ilya.kramt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
psycopg2-binary = "^2.9.1"
django-embed-video = "^1.4.0"
Pillow = "^8.4.0"
django-simple-captcha = "^0.5.14"
django-ckeditor = "^6.1.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
